[Refactor]: Rename Evento to Casilla, update position type to int, and create CasillaRepository for database interactions

+++ Pontiland/src/main/java/com/NullPtr/Pontiland/entities/ICasilla.java
@@ -8,7 +8,7 @@ public interface ICasilla {
 
   void setPosicionTablero(Byte posicionTablero);
 
-  byte getPosicionTablero();
+  int getPosicionTablero();
 
   void setNombreCasilla(String nombreCasilla);
